Refactor admin into package
