+++ transport/admin/__init__.py
@@ -1,0 +1,10 @@
+"""Admin package initialization."""
+
+from .common import *  # noqa
+from .vehicles import *  # noqa
+from .cte import *  # noqa
+from .mdfe import *  # noqa
+from .payments import *  # noqa
+from .config import *  # noqa
+from .dashboard import *  # noqa
+from .registry import *  # noqa
